Issue 43: initial wrapping of mogp `GaussianProcess` (#62)
* Merge 'iss46-change-terminology' into 'iss43-wrapping-mogp'

Squashed commit of the following:

commit c6bf5f2ae29809d98ee5ae781a8591f2a188a87a
Author: Thomas Hawes <105276609+thawes-rse@users.noreply.github.com>
Date:   Fri May 5 16:43:01 2023 +0100

    Improve test for immutable TrainingDatum fields

commit 8d9320bb52c24e055f18b95f88e4419005da910f
Author: Thomas Hawes <105276609+thawes-rse@users.noreply.github.com>
Date:   Fri May 5 16:38:02 2023 +0100

    Rename observation to (simulator) output

commit 012f54471e86d94bc2a416d637716ea0414d3b1a
Author: Thomas Hawes <105276609+thawes-rse@users.noreply.github.com>
Date:   Fri May 5 16:23:12 2023 +0100

    Rename experiment to input

* Remove unnecessary import

* Start class wrapping mogp emulators

* Add attribute to recover supplied GP

* Make gp a read-only property of MogpEmulator

* Add type hinting for GaussianProcess

* Add property for training data

Currently returns the data used to train the underlying GP. Also
includes class methods for Input and TrainingDatum to support the
creation of objects from Numpy arrays.

* Add input validation for from_array

* Add argument validation for Input constructor

* Refine error message about NaN values

* Add missing type hints

* Correct name and description of unit test

The test method in question was reusing the name of another test
method.

* Add validation of TrainingDatum output parameter

Have refactored the various checks on real numbers into functions
living in a new module concerned with validation.

* Encapsulate validation into a class

* Fix validation of observation in TrainingDatum

Before, an iterable of real numbers would have passed validation.

* Refactor validation out of classes

On reflection, classes are unnecessarily verbose for holding the
validation checks and no state needs to be tracked. So adopt a more
functional approach instead.

* Update docstrings and add missing type hints

* Add documentation for validation functions

Including type hinting.

* Add basic ability to fit a GP

Currently does not support fitting with supplied training data.
The AbstractEmulator abstract method fit() has also been updated
to now only optionally require training data be supplied to it.

* Fit while respecting hyperparameter bounds

Features an implementation of a hyperparameter estimation routine
that respects bounds on hyperparameters. This has been adapted from
the mogp-emulator package. See issue #58 for improvements still to
be done.

* Transform hyperparameters into raw params

The fit method now takes in bounds as defined on the hyperparameters
rather than the 'raw' hyperparameters, using the transformations
defined in:
https://mogp-emulator.readthedocs.io/en/latest/implementation/GPParams.html#mogp_emulator.GPParams.GPParams

* Add docstrings and type hinting to methods

* Update poetry dependencies

Now require numpy, scipy, mogp-emulator as core dependencies.

* Add copy of mogp-emulator fitting.py

This is in preparation for moving the mogp-adapted code into its
own module.

* Add bounds to optimisation for fitting functions

* Move fitting with bounds out of MogpEmulator fit

Also fix the helper function that asserts a value is approximately
within bounds.

* Update signature of AbstractEmulator fit method

Now includes the hyperparameter bounds parameter.

* Start implementing fit with training data

This starts adding the capability to fit an emulator to supplied
data. However, there is a bug in the current implementation, where
settings used to construct the underlying MOGP GP are not preserved.
This is captured in a failing unit test.

* Construct GaussianProcess as part of initialiser

MogpEmulator now takes in args and kwargs required by MOGP's
GuassianProcess class and constructs an object of this class,
while also storing the kwargs used in doing so. This allows us to
retrain the underlying GaussianProcess when fitting with new
training data.

* Remove licence notice

No longer needed as the code adapted from mogp-emulator now lives
in exauq.utilities.mogp_fitting, with licence notice there.

* Rename method

* Mild refactoring of fit method

Removes some nested if statements.

* Ignore inputs and targets in initialisation

MogpEmulators are now initialised with just the kwargs required for
constructing mogp GaussianProcess objects. Any inputs and targets
are ignored and replaced with empty data. This has simplified the
interface to the fit method.

* Remove training_data from AbstractEmulator

This may not be needed going forward and it would be best to
ensure this ABC only includes what is strictly necessary for working
with designers.

* Have MogpEmulator extend AbstractEmulator

* Update docstrings and type hinting

* Finish incomplete docstring

* Correct type hint

Also make other cosmetic changes.

* Add support for unconstrained bounds

* Add script demo-ing MogpEmulator wrapper

* Rewrite tests for bounds

These now use mocking to check directly that the fitting function
uses the correct bounds.

* Reformat files with Black

* Add module summary docstrings

* Tweak exception handling in class initialisation

In response to review, any non-Exception-derived exceptions will
be immediately raised as-is.

* Remove duplication in unit tests

As suggested in review.

* Make error messages more specific

As suggested in review.

* Remove invocation of superclass

As pointed out in review.

* Remove unnecessary import