Add previous simulations from log to simulator
If a simulations log file is provided to a simulator upon initialisation,
then it will gather any recorded simulations from there. Testing this
involvs mocking the SimulationsLog class with a simpler subclass that
returns pre-defined simulations. A function for creating such subclasses
has been added to the unit test file.