Return ouptut of previously computed simulations
If a previously seen input is supplied for computation, then the
corresponding simulation output is returned in the case where this
exists in the log file.

Currently this works by looking up against previously seen inputs
using strict equality comparison on the inputs. A version that uses
equality up to a tolerance should be implemented in the future.