Add AbstractGaussianProcess ABC
There are certain calculations with an emulator that assume it is
a Gaussian process. The new ABC allows for this distinction. As an
example, it gives a method for computing the normalised expected
square error at a training data point (the formula for which is
only valid if the emulator is a GP).