Remove assertion in covariance_matrix
We have made the choice to make this a 'bare bones' implementation.
This means that the default implementation should be efficient relative
to the implementation of the correlation method, and also doesn't
unduly make assumptions about the correlation method's behaviour in
edge cases, e.g. if the GP hasn't been trained on data.