Add improvements to manager script
This now creates a COMPLETED flag file as soon as the simulator script
finishes. This required wrapping the call to the simulator script in
a new 'runner' script, to support putting compound commands into the
background over SSH.

Also, the process group ID of this runner script is recorded, to
ensure that *all* subprocesses are killed upon job cancellation.