Modify API of simulator script
It is now expected that the simulator script required for
UnixServerScriptInterface will take in two command line args:
the path to the simulation input data CSV file and the path to
the simulation output file. This is expected to be simpler to
implement than the JSON config file method from before.