Improve job cancellation functionality
Now the cancellation job strategy will check the remote status
of the jobs to see if they need cancelling, thereby addressing
situations where the status of the job on the remote machine
and the status recorded in the simulations log are out of sync.

Also, the IDs of any job IDs that don't exist or that
have already terminated are displayed, along with a table of the
jobs that have been cancelled.