Add test for fltering training data correctly
Also includes a partial implementation which currently fails the test.
While a filtration from the top level to the bottom level should be
supplied in the input data, the output should not contain repeated
inputs across levels. If an input appears in level l, then it won't
appear in level l-1. At the bottom level only inputs that don't
feature in any of the higher levels are included.