Make PEI not dependent on input GP and domain
Upon initialisation of a PEICalculator instance, copies of `gp`
and `domain` are stored for subsequent calculations. This
removes the need update the internal state of the instance to
keep e.g. the repulsion points in line with any changes to the
training data for `gp`. The `gp` and `domain` properties have also
been removed to avoid changes to the internal state of PEICalculator.