Change update method input type to match fit input type.
Details:
-----------------------------------------------

The update method for single level GPs now sits
in AbstractGaussianProcess and is tested
equivalently through there.

The update method for MLGPs sits in the
MultiLevelGaussianProccess class where it
has to combine MultiLevel objects.

To do this combination the MultiLevel class
__add__ dunder method has been overwritten to
combine these allowing the use of the + operator
to combine MultiLevel objects.

These implementations also have new tests in the
appropriate modelling tests.